backend/requirements.txt
setuptools>=65.0.0
wheel>=0.37.0
Flask==3.0.3
Flask-CORS==4.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
python-dotenv==1.0.0
schedule==1.2.0
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
